backend/flask/requirements.txt
Flask
flask-cors
tensorflow
numpy
Pillow
PyPDF2
difflib
pymongo
requests 
python-dotenv
pymongo